retry on sqlite initialise
