+++ packages/db/src/base-sql.ts
@@ -1,10 +1,9 @@
-import { BlockEntry, Database, KeyValueEntry, defaultLogger } from '@acala-network/chopsticks-core'
-import { DataSource, QueryFailedError } from 'typeorm'
+import { BlockEntry, Database, KeyValueEntry } from '@acala-network/chopsticks-core'
+import { DataSource } from 'typeorm'
 import { HexString } from '@polkadot/util/types'
 
 import { BlockEntity, KeyValuePair } from './db/entities'
-
-const logger = defaultLogger.child({ name: 'base-sql' })
+import { retry } from './retry'
 
 function Retryable<T>(
   _target: any,
@@ -18,28 +17,6 @@ function Retryable<T>(
   }
 
   return descriptor
-}
-
-export async function retry<T>(fn: () => Promise<T>, maxRetries: number = 3, delay: number = 500): Promise<T> {
-  let retries = 0
-  while (retries < maxRetries) {
-    try {
-      return await fn()
-    } catch (error) {
-      if (error instanceof QueryFailedError) {
-        if (error.message.includes('SQLITE_BUSY')) {
-          retries++
-          logger.info(`SQLite is busy. Retrying in ${delay}ms (Attempt ${retries})...`)
-          await new Promise((r) => setTimeout(r, delay))
-        } else {
-          throw error
-        }
-      } else {
-        throw error
-      }
-    }
-  }
-  throw new Error(`Exceeded maximum retries (${maxRetries}) for SQLite busy error.`)
 }
 
 export abstract class BaseSqlDatabase implements Database {
