+++ packages/db/src/db/index.ts
@@ -1,6 +1,7 @@
 import { DataSource } from 'typeorm'
 
 import * as entities from './entities'
+import { retry } from '../retry'
 
 export const openDb = async (dbPath: string): Promise<DataSource> => {
   const source = new DataSource({
@@ -11,7 +12,7 @@ export const openDb = async (dbPath: string): Promise<DataSource> => {
     logging: false,
   })
 
-  await source.initialize()
+  await retry(() => source.initialize(), 3, 1000)
 
   return source
 }
